pyproject.toml
[tool.poetry]
name = "chgjx"
version = "0.1.0"
description = ""
authors = ["kevin <ye@126.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tabulate = "^0.9.0"
mysql-connector-python = "^9.0.0"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
